requirements.txt
Django==2.0.7
dj-database-url==0.5.0
dj-static==0.0.6
django-extensions==2.1.0
django-localflavor==2.0
django-widget-tweaks==1.4.2
Pillow==5.2.0
python-decouple==3.1
